chore: move account utils into separate accounts file
