+++ src/components/AccountDropdown/AccountDropdown.tsx
@@ -31,7 +31,7 @@ import { useTranslate } from 'react-polyglot'
 import { useSelector } from 'react-redux'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
-import { isValidAccountNumber } from 'lib/utils'
+import { isValidAccountNumber } from 'lib/utils/accounts'
 import { type ReduxState } from 'state/reducer'
 import { accountIdToLabel } from 'state/slices/portfolioSlice/utils'
 import {
