+++ src/components/AssetHeader/AccountLabel.tsx
@@ -1,7 +1,7 @@
 import { HStack, Tag, TagLabel } from '@chakra-ui/react'
 import type { AccountId } from '@shapeshiftoss/caip'
 import { RawText } from 'components/Text'
-import { accountIdToChainDisplayName } from 'lib/utils'
+import { accountIdToChainDisplayName } from 'lib/utils/accounts'
 import { accountIdToLabel } from 'state/slices/portfolioSlice/utils'
 
 type AccountLabelProps = { accountId: AccountId }
