+++ src/components/Layout/Header/DegradedStateBanner.tsx
@@ -21,7 +21,7 @@ import { useTranslate } from 'react-polyglot'
 import { useSelector } from 'react-redux'
 import { LazyLoadAvatar } from 'components/LazyLoadAvatar'
 import { RawText, Text } from 'components/Text'
-import { accountIdToFeeAssetId } from 'lib/utils'
+import { accountIdToFeeAssetId } from 'lib/utils/accounts'
 import { portfolioApi } from 'state/slices/portfolioSlice/portfolioSlice'
 import { selectAssets, selectPortfolioErroredAccountIds } from 'state/slices/selectors'
 import { useAppDispatch } from 'state/store'
