+++ src/features/defi/providers/fox-farming/hooks/useFoxFarming.ts
@@ -8,7 +8,7 @@ import { encodeFunctionData, getAddress, maxUint256 } from 'viem'
 import { useFoxEth } from 'context/FoxEthProvider/FoxEthProvider'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { toBaseUnit } from 'lib/math'
-import { isValidAccountNumber } from 'lib/utils'
+import { isValidAccountNumber } from 'lib/utils/accounts'
 import {
   assertGetEvmChainAdapter,
   buildAndBroadcast,
