+++ src/pages/Accounts/AccountTxHistory.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom'
 import { AssetHeader } from 'components/AssetHeader/AssetHeader'
 import { Main } from 'components/Layout/Main'
 import { AssetTransactionHistory } from 'components/TransactionHistory/AssetTransactionHistory'
-import { accountIdToFeeAssetId } from 'lib/utils'
+import { accountIdToFeeAssetId } from 'lib/utils/accounts'
 import { selectAssetById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
