+++ src/pages/Dashboard/components/ErroredTxHistoryAccounts.tsx
@@ -16,7 +16,7 @@ import { useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { ChainIcon } from 'components/ChainMenu'
 import { RawText, Text } from 'components/Text'
-import { getAccountTitle } from 'lib/utils'
+import { getAccountTitle } from 'lib/utils/accounts'
 import { isUtxoChainId } from 'lib/utils/utxo'
 import {
   selectAccountNumberByAccountId,
