+++ src/plugins/walletConnectToDapps/components/header/AddressLinks.tsx
@@ -4,7 +4,8 @@ import { fromAccountId } from '@shapeshiftoss/caip'
 import { useMemo } from 'react'
 import { MiddleEllipsis } from 'components/MiddleEllipsis/MiddleEllipsis'
 import { Text } from 'components/Text'
-import { accountIdToFeeAssetId, isSome } from 'lib/utils'
+import { isSome } from 'lib/utils'
+import { accountIdToFeeAssetId } from 'lib/utils/accounts'
 import { selectAssetById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
