+++ src/state/slices/opportunitiesSlice/resolvers/cosmosSdk/index.ts
@@ -1,6 +1,7 @@
 import { cosmosChainId, fromAccountId } from '@shapeshiftoss/caip'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
-import { accountIdToFeeAssetId, isFulfilled, isRejected, isSome } from 'lib/utils'
+import { isFulfilled, isRejected, isSome } from 'lib/utils'
+import { accountIdToFeeAssetId } from 'lib/utils/accounts'
 import { assertGetCosmosSdkChainAdapter } from 'lib/utils/cosmosSdk'
 import type { ReduxState } from 'state/reducer'
 import { selectAssetById } from 'state/slices/assetsSlice/selectors'
