+++ src/state/slices/portfolioSlice/selectors.ts
@@ -23,7 +23,8 @@ import { isMobile } from 'lib/globals'
 import { fromBaseUnit } from 'lib/math'
 import { getMaybeCompositeAssetSymbol } from 'lib/mixpanel/helpers'
 import type { AnonymizedPortfolio } from 'lib/mixpanel/types'
-import { hashCode, isValidAccountNumber } from 'lib/utils'
+import { hashCode } from 'lib/utils'
+import { isValidAccountNumber } from 'lib/utils/accounts'
 import type { ReduxState } from 'state/reducer'
 import { createDeepEqualOutputSelector } from 'state/selector-utils'
 import {
